Change to memcpy

+++ examples/server/server.c
@@ -2404,11 +2404,11 @@ THREAD_RETURN WOLFSSL_THREAD server_test(void* args)
             /* Write data */
             if (!useWebServerMsg) {
                 write_msg = msg;
-                write_msg_sz = XSTRLEN(msg);
+                write_msg_sz = (int) XSTRLEN(msg);
             }
             else {
                 write_msg = webServerMsg;
-                write_msg_sz = XSTRLEN(webServerMsg);
+                write_msg_sz = (int) XSTRLEN(webServerMsg);
             }
             ServerWrite(ssl, write_msg, write_msg_sz);
